Define OpenAIResponse Class & Deserialize
Defined a new class called OpenAIResponse so that the json response from the
API can be deserialized by Unity's JsonUtility.FromJson method. The text response
from the LLM is stored as an attribute in the AIRequest class.
OpenAITest has also been renamed to AIRequest to better reflect its use in the
code

Issue Number: #286