Changed ChatGPT answer to use WitAI TTS temporarily
For now, ChatGPT answer is given with previous WitAI TTS to align with scripted dialogue. This is a temporary fix to align the voices while we work out how to change scripted dialogue to use OpenAI TTS. It is probably a good option to have for the organization to switch between the two anyway.

Issue Number: #319