Fix and improve various things
One of the dialogue texts received a bad request error when being sent to OpenAI TTS due to a newline character. This is now fixed. When using OpenAI, there is now a slight delay before the talking animation begins due to this solution taking slightly longer to generate speech. Additionally, "isListening" variable was not saved correctly.