Fix listening animation
The listening animation no longer plays on repeat. The animation transition was
missing a condition. The condition of _isListeningHash = true is now required
for the animation to play. The animation trigger has been moved from the
DisplayThinking coroutine to when the player presses E. This has to be
implemented for the VR headset as well

Issue Number: #343